+++ js/src/modals/CreateAccount/createAccount.js
@@ -34,8 +34,8 @@ import RawKey from './RawKey';
 import RecoveryPhrase from './RecoveryPhrase';
 
 import { createIdentityImg } from '../../api/util/identity';
-import print from './AccountDetails/print';
-import recoveryPage from './AccountDetails/recovery-page.ejs';
+import print from './print';
+import recoveryPage from './recovery-page.ejs';
 import ParityLogo from '../../../assets/images/parity-logo-black-no-text.svg';
 
 const TITLES = {
