+++ js/src/modals/CreateAccount/recovery-page.ejs
